+++ opencv/src/main/java/org/bytedeco/opencv/presets/opencv_bgsegm.java
@@ -36,7 +36,7 @@ import org.bytedeco.javacpp.tools.InfoMapper;
             link = "opencv_bgsegm@.4.5"
         ),
         @Platform(value = "ios", preload = "libopencv_bgsegm"),
-        @Platform(value = "windows", link = "opencv_bgsegm453")
+        @Platform(value = "windows", link = "opencv_bgsegm454")
     },
     target = "org.bytedeco.opencv.opencv_bgsegm",
     global = "org.bytedeco.opencv.global.opencv_bgsegm"